Adding  authorization to the listings controlled by owners

+++ Backend/src/routes/listingRoutes.js
@@ -93,7 +93,21 @@ router.delete(
   "/:id",
   LoggedIn,
   wrapAsync(async (req, res) => {
+    
     const { id } = req.params;
+
+    const listing = await Listing.findById(id) // extract listing from the database
+    
+    // 2. CHECK LOGIC: req.user._id ko listing.Owner se compare karein
+    if(!listing.Owner.equals(req.user._id)){
+      return res.status(403).json({
+        success:false,
+        message:"Your are not authorized"
+      })
+    }
+
+    // Mongoose ObjectIDs ko compare karne ke liye '.equals()' use hota hai
+    
     await Listing.findByIdAndDelete(id);
     res.json({ success: true, message: "Deleted successfully" });
   })
